Configure logging
Logging in JSON would make more sense in a true production environment
where we were aggregating logs and so forth, but since this is likely a
one-off and other students will be working on this with me, the more
human-friendly we can make things, the better.

Hence, we're pretty-printing, enabling debug logs, and getting rid of
the stuff where it says what host/process is running.